+++ ToolbarIcons/CHANGELOG.md
@@ -1,8 +1,21 @@
 # Toolbar Icons Change Log
 
-## 2.0.1 (Unreleased)
+## 2.1.0 (Unreleased)
 
+* Added a default background to all icons.
 * Added alert if FuryCore is installed.
+* Added direct integration with some mods.
+    * Always Scroll Map
+    * Chests Anywhere
+    * CJB Cheats Menu
+    * CJB Item Spawner
+    * Data Layers
+    * Debug Mode
+    * Dynamic Game Assets
+    * Horse Flute Anywhere
+    * Instant Buildings
+    * Lookup Anything
+    * Stardew Aquarium
 
 ## 2.0.0 (July 2, 2022)