+++ ToolbarIconsCP/CHANGELOG.md
@@ -1,5 +1,20 @@
 ﻿# Toolbar Icons for FuryCore Change Log
 
+## 2.1.0 (Unreleased)
+
+* Removed icons for mods that are directly integrated into ToolbarIcons.
+  * Always Scroll Map
+  * Chests Anywhere
+  * CJB Cheats Menu
+  * CJB Item Spawner
+  * Data Layers
+  * Debug Mode
+  * Dynamic Game Assets
+  * Horse Flute Anywhere
+  * Instant Buildings
+  * Lookup Anything
+  * Stardew Aquarium
+
 ## 2.0.0 (July 1, 2022)
 
 * Changed dependency to ToolbarIcons